Move API middleware to class, remove version number from config
