+++ backend/src/config.php
@@ -4,8 +4,6 @@ return [
     'config' => [
         'soundboard' => [
             'sampleBaseDirectory' => 'samples',
-            // TODO: Dedupe.
-            'versionNumber' => '3.4.1',
         ],
     ],
 ];
